- Bing Webmaster Tools integration
- Import Bing Search Keywords
- Query parameter access to Google Search Analytics API via
data-capture-run.php file
- Ability to return API data instead of importing to DB via
data-capture-run.php file
- Comment and code cleanup
- Upgrade script
- Bug fix for CTR
- Settings layout modified to accommodate Bing Webmaster Tools

+++ organic-search-analytics.sql
@@ -36,7 +36,7 @@ CREATE TABLE IF NOT EXISTS `search_analytics` (
   `query` varchar(500) NOT NULL,
   `impressions` int(11) NOT NULL,
   `clicks` int(11) NOT NULL,
-  `ctr` int(11) NOT NULL,
+  `ctr` float NOT NULL,
   `avg_position` int(11) NOT NULL,
   `avg_position_click` int(11) NOT NULL,
   PRIMARY KEY (`id`)
